requirements.txt
# This file represents the needs for the deployer script to function, while the
# dev-requirements.txt file represents the needs in this repo in general.
#

# ruamel.yaml is used to read and write .yaml files.
ruamel.yaml==0.17.21

# auth0 is used to communicate with Auth0's REST API that we integrate with in
# various ways.
auth0-python==3.24.0

# jsonschema is used for validating cluster.yaml configurations
jsonschema==4.17.0

# rich and py-markdown-table are used for pretty printing outputs that would otherwise
# be difficult to parse by a human
rich==13.0.0
py-markdown-table==0.3.1

# jhub_client, pytest, and pytest-asyncio are used for our health checks
jhub-client==0.1.6
pytest==7.2.0
pytest-asyncio>=0.17

# Used to generate templates
jinja2==3.1.2

# Used for the debug CLI
typer==0.7.0
escapism==1.0.1